% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allwidths_lake.R
\name{allwidths_lake}
\alias{allwidths_lake}
\title{Toutes les droites orthogonales à la ligne de base -longueur maximale- équidistant de la distance en m}
\usage{
allwidths_lake(sppolygon, linemax, distance)
}
\arguments{
\item{sppolygon}{un polygone spatialisé -objet sf- masse d'eau}

\item{linemax}{longueur la plus grande sur la masse d'eau -objet sf-}

\item{distance}{intervalle entre deux points le long de la longueuer maximale de la masse d'eau -distance en m-}
}
\value{

}
\description{
Toutes les droites orthogonales à la ligne de base -longueur maximale- équidistant de la distance en m
}
\examples{
library(lakemetrics)
data(lac)
linemax <- linemax_lake(sppolygon=lac)
allwidths_lake(sppolygon=lac, linemax=linemax, distance=100)

}
