% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_lake.R
\name{area_lake}
\alias{area_lake}
\title{Surface en eau d'un lac}
\usage{
area_lake(sppolygon)
}
\arguments{
\item{sppolygon}{un polygone spatialisé -objet sf- masse d'eau}
}
\value{
valeur de la surface en eau en m²
}
\description{
Surface en eau d'un lac
}
