% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_lakes.R
\name{class_lakes}
\alias{class_lakes}
\title{Classe dans l'ordre décroissant les lacs d'un même ensemble d'une masse d'eau selon la surface en eau}
\usage{
class_lakes(sppolygon)
}
\arguments{
\item{sppolygon}{un objet sf correspondant un polygone de la masse d'eau}
}
\value{
le même polygone spatialisé avec deux variables en plus la surface en eau -area- et une variable ordinale qui classe dans l'ordre décroissant les lacs selon leurs surface.
}
\description{
Classe dans l'ordre décroissant les lacs d'un même ensemble d'une masse d'eau selon la surface en eau
}
