% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthwavebase_lake.R
\name{depthwavebase_lake}
\alias{depthwavebase_lake}
\title{Profondeur d’influence des vagues (Depth of Wave Base)}
\usage{
depthwavebase_lake(area)
}
\arguments{
\item{area}{aire de la masse d'eau en km²}
}
\value{
Cette valeur correspond à la profondeur maximale d’influence de l’énergie des vagues sur les sédiments de fond pour une surface de plan d’eau (Hakanson 2005). Ce paramètre doit servir à mieux décrire une partie du phénomène de remise en suspension des sédiments qui peut avoir des effets sur la physico-chimie du plan d’eau, la chaine trophique ou encore la répartition de la faune  et/ou la flore. Exprimée en mètre (m)
}
\description{
Profondeur d’influence des vagues (Depth of Wave Base)
}
\examples{
library(lakemetrics)
depthwavebase_lake(area=50000)

}
\references{
Hakanson 2005
}
