% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffrasterpts.R
\name{diffrasterpts}
\alias{diffrasterpts}
\title{Différence de profondeur entre les pts_bathy et le modèle}
\usage{
diffrasterpts(depth_raster, pts_bathy)
}
\arguments{
\item{depth_raster}{raster XY value : profondeur}

\item{pts_bathy}{une collection de points -objet sf- de dimension XYZ  ou Z : profondeur}
}
\value{

}
\description{
Différence de profondeur entre les pts_bathy et le modèle
}
\examples{
library(lakemetrics)

}
