% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lac.R
\docType{data}
\name{lac}
\alias{lac}
\title{Un objet sf polygone de la masse d'eau}
\format{
Polygon sf
\describe{
  \item{epsg}{2154}
}
}
\source{
\url{Sandre}
}
\usage{
lac
}
\description{
Un objet sf polygone de la masse d'eau
}
\keyword{datasets}
