% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferlakeinbv_lake.R
\name{lakebufferinbv_lake}
\alias{lakebufferinbv_lake}
\title{Création d'une zone tampon autour de la masse d'eau en prenant en compte le découpage du Bassin versant.}
\usage{
lakebufferinbv_lake(sppolygon, bv, distance)
}
\arguments{
\item{sppolygon}{un objet sf correspondant un polygone de la masse d'eau}

\item{bv}{polygone du bassin versant de la masse d'eau -un objet sf-}

\item{distance}{distance de la zone tampon valeur valeur numérique en m}
}
\value{
un polygone -objet sf- de la zone tampon
}
\description{
Création d'une zone tampon autour de la masse d'eau en prenant en compte le découpage du Bassin versant.
}
\examples{
library(lakemetrics)
data(lac)
}
