% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthmax_allpoints_lake.R
\name{lengthmax_allpoints_lake}
\alias{lengthmax_allpoints_lake}
\title{Comparaison des longueurs pour chaque pont afin de déterminer la longueur maximale de la masse d'eau}
\usage{
lengthmax_allpoints_lake(linemaxpoints)
}
\arguments{
\item{linemaxpoints}{un objet sf -linestring- ensemble de longueur maximale pour chaque points de la masse d'eau}
}
\value{
la longueur maximale de la masse d'eau en obejt sf -linestring-
}
\description{
Comparaison des longueurs pour chaque pont afin de déterminer la longueur maximale de la masse d'eau
}
\examples{
library(lakemetrics)
}
