% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemax_lake.R
\name{linemax_lake}
\alias{linemax_lake}
\title{Longueur la plus grande sur la masse d'eau}
\usage{
linemax_lake(sppolygon)
}
\arguments{
\item{sppolygon}{un objet sf correspondant au polygone de la masse d'eau}
}
\value{
la longueur la plus grande sur la masse d'eau -objet sf-
}
\description{
Longueur la plus grande sur la masse d'eau
}
\examples{
library(lakemetrics)
data(lac)
linemax_lake(lac)

}
