% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemaxpoint_lake.R
\name{linemaxpoint_lake}
\alias{linemaxpoint_lake}
\title{Longueur maximale pour un point spatialisé de la masse d'eau}
\usage{
linemaxpoint_lake(point_sf, points_lake)
}
\arguments{
\item{point_sf}{un objet sf correspondant à un point spatialisé de la masse d'eau}

\item{points_lake}{un objet sf -mulitpoint- correspondant à tout les points définissant  la masse d'eau}
}
\value{

}
\description{
Longueur maximale pour un point spatialisé de la masse d'eau
}
\examples{
library(lakemetrics)

}
