% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineperdendicular.R
\name{lineperdendicular}
\alias{lineperdendicular}
\title{Ligne perpendiculaire selon un point sur l}
\usage{
lineperdendicular(sppolygon, linemax, X, Y)
}
\arguments{
\item{sppolygon}{un polygone spatialisé -objet sf- masse d'eau}

\item{linemax}{objet sf correspondant à la longueur maximale sur la masse d'eau -linestring-}

\item{coord}{coordonnées du point d'intersection sur la longueur maximale de la masse d'eau}
}
\value{

}
\description{
Ligne perpendiculaire selon un point sur l
}
\examples{
library(lakemetrics)

}
