% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanwidthcalc_lake.R
\name{meanwidthcalc_lake}
\alias{meanwidthcalc_lake}
\title{Largeur moyenne calculée selon l'aire et la longueur maximale de la masse d'eau}
\usage{
meanwidthcalc_lake(sppolygon)
}
\arguments{
\item{sppolygon}{un polygone spatialisé -objet sf- masse d'eau}
}
\value{
Valeur de la largeur moyenne de la masse d'eau selon la formule
}
\description{
Largeur moyenne calculée selon l'aire et la longueur maximale de la masse d'eau
}
\examples{
library(lakemetrics)

}
