% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntbl_ibml.R
\name{ntb_ibml}
\alias{ntb_ibml}
\title{Nombre de transects à retenir selon la nome XP T90-328}
\usage{
ntb_ibml(sppolygon)
}
\arguments{
\item{sppolygon}{un objet sf correspondant un polygone de la masse d'eau}
}
\value{
le nombre de transects à tracer le long de la ligne de base -valeur numérique-
}
\description{
Nombre de transects à retenir selon la nome XP T90-328
}
\examples{
library(lakemetrics)
ntb_ibml(lac)

}
