% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_linemax.R
\name{points_linemax}
\alias{points_linemax}
\title{Répartition des points équidistants le long de longueur maximale de la masse d'eau}
\usage{
points_linemax(linemax, distance)
}
\arguments{
\item{linemax}{objet sf correspondant à la longueur maximale sur la masse d'eau -linestring-}

\item{distance}{intervalle entre deux points le long de la longueuer maximale de la masse d'eau -distance en m-}
}
\value{

}
\description{
Répartition des points équidistants le long de longueur maximale de la masse d'eau
}
\examples{
library(lakemetrics)

}
