% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi_lake.R
\name{sdi_lake}
\alias{sdi_lake}
\title{Indice de complexité des rives -SDI-}
\usage{
sdi_lake(sppolygon)
}
\arguments{
\item{sppolygon}{un polygone spatialisé -objet sf- masse d'eau}
}
\value{
indice de complexité des rives
}
\description{
Indice de complexité des rives -SDI-
}
\examples{
library(lakemetrics)

}
