% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoreline.R
\name{shoreline_lake}
\alias{shoreline_lake}
\title{Périmètre de la masse d'eau}
\usage{
shoreline_lake(sppolygon)
}
\arguments{
\item{sppolygon}{un objet sf correspondant un polygone de la masse d'eau}
}
\value{
distance du périmètre en m -numeric-
}
\description{
Périmètre de la masse d'eau
}
\examples{
library(lakemetrics)
data(lac)
shoreline_lake(lac)

}
