% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects_ibml.R
\name{transects_ibml}
\alias{transects_ibml}
\title{Transects et la ligne de base  conformément au protocole d'échantillonnage}
\usage{
transects_ibml(sppolygon, linemax)
}
\arguments{
\item{sppolygon}{un polygone spatialisé -objet sf- masse d'eau}

\item{linemax}{la longueur la plus grande sur la masse d'eau -objet sf-}
}
\value{
ligne de base et les transects perpendiculaires -objet sf-
}
\description{
Transects et la ligne de base  conformément au protocole d'échantillonnage
}
\examples{
library(lakemetrics)
library(ggplot2)
data(lac)
baseline <- linemax_lake(lac)
transects <- transects_ibml(lac,baseline)
transects

}
