% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widthmax_lake.R
\name{widthmax_lake}
\alias{widthmax_lake}
\title{Largeur maximale de la masse d'eau selon la ligne de base -longueur maximale-}
\usage{
widthmax_lake(sppolygon, linemax, distance)
}
\arguments{
\item{sppolygon}{un polygone spatialisé -objet sf- masse d'eau}

\item{linemax}{longueur la plus grande sur la masse d'eau -objet sf-}

\item{distance}{intervalle entre deux points le long de la longueuer maximale de la masse d'eau -distance en m-}
}
\value{

}
\description{
Largeur maximale de la masse d'eau selon la ligne de base -longueur maximale-
}
\examples{
library(lakemetrics)

}
