% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathymetry_lake.R
\name{bathymetry_lake}
\alias{bathymetry_lake}
\title{Bathymmétrie de la masse d'eau au format raster}
\usage{
bathymetry_lake(sppolygon, pts_bathy)
}
\arguments{
\item{sppolygon}{un objet sf correspondant un polygone de la masse d'eau}

\item{pts_bathy}{une collection de points -objet sf- de dimension XYZ  ou Z : profondeur}
}
\value{
Bathymmétrie de la masse d'eau au format raster ; plusieurs algorithmes pour la création de la bathymétrie sont testés.
\describe{
  \item{IDW} {Inverse Distance Weighting}
  \item{TPRS} {Thin Plate Regression Spline}
  \item{} {Soap Film Smooth}
}

Le choix se fait juste par l'écart le plus petit entre les pts_bathy et le modèle -en valeur absolue-.
}
\description{
Bathymmétrie de la masse d'eau au format raster
}
\examples{
library(lakemetrics)

}
\references{
https://fishandwhistle.net/post/2019/bathymetry-lake-volume-estimation-using-r/
}
